Fixed some linux compile warnings with dtDirectorQt

+++ trunk/delta3d/src/dtDirectorQt/undomanager.cpp
@@ -46,7 +46,7 @@ namespace dtDirector
    {
       if (mModifyIndex == -1) return true;
 
-      if (mModifyIndex != mUndoEvents.size()) return true;
+      if (mModifyIndex != (int)mUndoEvents.size()) return true;
 
       return false;
    }
